requirements.txt
# DistroWiki API - Dependências Python

# ==============================================================================
# Core Framework
# ==============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# ==============================================================================
# Database & Auth
# ==============================================================================
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0

# ==============================================================================
# HTTP Client
# ==============================================================================
httpx>=0.25.0

# ==============================================================================
# Web Scraping (opcional - se usar DistroWatch)
# ==============================================================================
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ==============================================================================
# Validação e Parsing
# ==============================================================================
python-multipart>=0.0.6
email-validator>=2.0.0

# ==============================================================================
# Variáveis de Ambiente
# ==============================================================================
python-dotenv>=1.0.0

# ==============================================================================
# Logging
# ==============================================================================
python-json-logger>=2.0.0

# ==============================================================================
# IA - Perplexity API (Sonar Reasoning Pro)
# ==============================================================================
openai>=1.0.0

# ==============================================================================
# Google Sheets API com OAuth 2.0
# ==============================================================================
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Dependências adicionais do Google (instaladas automaticamente, mas explicitadas)
googleapis-common-protos>=1.56.0
protobuf>=3.19.0,<5.0.0
cachetools>=4.2.0

# ==============================================================================
# Cache (Redis via Upstash)
# ==============================================================================
upstash-redis>=0.16.0
